Alter behavior of asyncrun with python scripts
Note: not working with ruby scripts as for some reason vim-polyglot
doesn't set makeprg variable properly.

Vim-polyglot sets makeprg to either 'make' or 'python' depending on the
presence of a Makefile. This makes it easier for me to control how I
want to quickly prototype a Python script in Vim.